Add back-up functionality
Functions:
* Back-up original recording locally to disk and to backblaze b2, with
  `__original` appended to the filename.
    E.g., interview-1__original.garageband
* Back-up final edit to backblaze b2, with timestamp appended to
  filename.
    E.g., interview-1-2024-08-14T10:00:00.000Z.garageband
* Back-up episode MP3 audio file to backblaze b2
    E.g., 'Episode 1: It Begins.MP3'
* Back-up any file
    E.g., note.txt

For back-ups to Backblaze cloud storage, a file is first zipped and then
the zip is uploaded to Backblaze. This saved ~ 200 MB / file in test
uploads of ~ 1 GB.

The user's Backblaze API key info, bucket name, and author name are
saved in UserDefaults, via the application preferences.
If the data is not found in UserDefaults on application launch and on
attempted use, then the application preferences window is shown so the
user can provide the data. Validation and feedback TBD.

Success and failure feedback messages are provided to the user.
Retrying on failure is left to the user. Automatic retrying of
applicable failures TBD.